Add user authentication service and repository
A new user authentication service and repository have been added to the project. The AuthService includes a method for validating a user by ID, if the user does not exist, an exception is thrown. AuthRepository is utilizing the new BaseRepository for common database operations. Alongside this addition, repository interfaces and their implementations have also been created for IRepository, IUnitOfWork and their BaseEntity. This forms the foundation for our repository pattern implementation and will provide the necessary tools for managing and isolating database operations.

The UserEntity model in the Shared project has been altered to include 'Username', 'Email', 'PasswordHash' and two timestamps 'CreatedAt' and 'UpdatedAt'. These fields will be necessary for further implementation of user management and authentication.

The codebase changes also reflect a shift towards Dependency Injection for databse operations with the introduction of DbContext and its usages across modified files. The necessary configurations for DI are added in the Startup.cs file for Jwt token service and AuthRepository.

In the .csproj files for both Shared and Auth, the necessary package references for the EntityFrameworkCore and other dependencies have been added.

The solution toward adopting Repository Pattern would provide a cleaner, more testable project architecture.