Add Feed service to TournaBet project
A new service called Feed has been added to the TournaBet project. This service contributes to the overall functionality of the application by implementing a Startup class, a FeedDbContext class, and the associated files necessary for these changes to run correctly, such as a .csproj and an appsettings.json file. A reference to the newly created Feed project has also been included in the TournaBet.sln file. Changes were made to the Shared.csproj file to support these integrations.